nvme: simplify dma/cmb mappings
Instead of handling both QSGs and IOVs in multiple places, simply use
QSGs everywhere by assuming that the request does not involve the
controller memory buffer (CMB). If the request is found to involve the
CMB, convert the QSG to an IOV and issue the I/O. The QSG is converted
to an IOV by the dma helpers anyway, so the CMB path is not unfairly
affected by this simplifying change.

As a side-effect, this patch also allows PRP/SGLs to be located in the
CMB. The logic ensures that if some of the PRP/SGL is in the CMB, all of
it must be located there.

Signed-off-by: Klaus Birkelund Jensen <klaus.jensen@cnexlabs.com>