nvme: add an OpenChannel 2.0 NVMe device (ocssd)
This adds a new 'ocssd' block device that emulates an OpenChannel 2.0
device. The device is backed by a new 'ocssd' block backend that is
based on the raw format driver but includes a header that holds the
device geometry and write data requirements. This new block backend is
special in that the size is not specified explicitly but in terms of
sector size, number of chunks, number of parallel units, etc. This
called for the addition of the `no_size_required` field in `struct
BlockDriver` to not fail image creation when the size parameter is
missing.

The ocssd block device is an individual device but shares a lot of
code with the nvme device. Thus, some core functionality of nvme/nvme.c
has been exported for use by nvme/ocssd.c.

Thank you to the following people for their contributions to the
original qemu-nvme (github.com/OpenChannelSSD/qemu-nvme) implementation
from which this code was originally derived.

  Matias Bjørling <mb@lightnvm.io>
  Javier González <javier@javigon.com>
  Simon Andreas Frimann Lund <ocssd@safl.dk>
  Hans Holmberg <hans@owltronix.com>
  Jesper Devantier <contact@pseudonymous.me>

Signed-off-by: Klaus Birkelund Jensen <klaus.jensen@cnexlabs.com>